fix: misc small bugs from code review (#180)
- calculateKcalPerPackage returns null instead of 0.0 when data is unavailable
- fix PantryControllerTest and PantryServiceTest to match updated consumeItem signature
- fix PantryServiceTest constructor to include LocalDatasetLookupService and LocalDatasetProductMapper mocks
- fix DailyNutrientIntakeServiceTest ambiguous null cast

+++ src/test/java/ch/uzh/ifi/hase/soprafs26/service/DailyNutrientIntakeServiceTest.java
@@ -136,7 +136,7 @@ class DailyNutrientIntakeServiceTest {
         assertNull(service.recordConsumedPantryItem(null, pantryItem(1L), 1, Instant.now()));
         assertNull(service.recordConsumedPantryItem(1L, null, 1, Instant.now()));
         assertNull(service.recordConsumedPantryItem(1L, pantryItem(null), 1, Instant.now()));
-        assertNull(service.recordConsumedPantryItem(1L, pantryItem(1L), null, Instant.now()));
+        assertNull(service.recordConsumedPantryItem(1L, pantryItem(1L), (Integer) null, Instant.now()));
         assertNull(service.recordConsumedPantryItem(1L, pantryItem(1L), 0, Instant.now()));
 
         when(micronutrientsRepository.findByPantryItemId(1L)).thenReturn(Optional.empty());
